refactor(processor): extract page selection
Move page-range resolution and render-region bounds checks into a processor helper so processDocument stays focused on orchestration.
